+++ src/main/java/com/algaworks/algafood/api/controller/CategoryController.java
@@ -1,7 +1,7 @@
 package com.algaworks.algafood.api.controller;
 
-import com.algaworks.algafood.api.model.CategoryRequest;
-import com.algaworks.algafood.api.model.CategoryResponse;
+import com.algaworks.algafood.api.model.request.CategoryRequest;
+import com.algaworks.algafood.api.model.response.CategoryResponse;
 import com.algaworks.algafood.domain.service.CategoryService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
