+++ src/main/java/com/algaworks/algafood/api/controller/CityController.java
@@ -1,7 +1,7 @@
 package com.algaworks.algafood.api.controller;
 
-import com.algaworks.algafood.api.model.CityRequest;
-import com.algaworks.algafood.api.model.CityResponse;
+import com.algaworks.algafood.api.model.request.CityRequest;
+import com.algaworks.algafood.api.model.response.CityResponse;
 import com.algaworks.algafood.domain.service.CityService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
