+++ src/main/java/com/algaworks/algafood/api/controller/PaymentMethodController.java
@@ -1,7 +1,7 @@
 package com.algaworks.algafood.api.controller;
 
-import com.algaworks.algafood.api.model.PaymentMethodRequest;
-import com.algaworks.algafood.api.model.PaymentMethodResponse;
+import com.algaworks.algafood.api.model.request.PaymentMethodRequest;
+import com.algaworks.algafood.api.model.response.PaymentMethodResponse;
 import com.algaworks.algafood.domain.service.PaymentMethodService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
