+++ src/main/java/com/algaworks/algafood/api/model/request/ProfileRequest.java
@@ -1,11 +1,11 @@
-package com.algaworks.algafood.api.model;
+package com.algaworks.algafood.api.model.request;
 
 import lombok.Data;
 
 import javax.validation.constraints.NotBlank;
 
 @Data
-public class StateRequest {
+public class ProfileRequest {
 
     @NotBlank
     private String name;
