+++ src/main/java/com/algaworks/algafood/api/model/request/RestaurantRequest.java
@@ -1,4 +1,4 @@
-package com.algaworks.algafood.api.model;
+package com.algaworks.algafood.api.model.request;
 
 import lombok.Data;
