+++ src/main/java/com/algaworks/algafood/api/model/request/StateIdRequest.java
@@ -1,11 +1,11 @@
-package com.algaworks.algafood.api.model;
+package com.algaworks.algafood.api.model.request;
 
 import lombok.Data;
 
 import javax.validation.constraints.NotNull;
 
 @Data
-public class CityIdRequest {
+public class StateIdRequest {
 
     @NotNull
     private Long id;
