+++ src/main/java/com/algaworks/algafood/api/model/request/StateRequest.java
@@ -1,0 +1,12 @@
+package com.algaworks.algafood.api.model.request;
+
+import lombok.Data;
+
+import javax.validation.constraints.NotBlank;
+
+@Data
+public class StateRequest {
+
+    @NotBlank
+    private String name;
+}
