+++ src/main/java/com/algaworks/algafood/api/model/response/CityModelResponse.java
@@ -1,4 +1,4 @@
-package com.algaworks.algafood.api.model;
+package com.algaworks.algafood.api.model.response;
 
 import lombok.Data;
