+++ src/main/java/com/algaworks/algafood/api/model/response/ProfileResponse.java
@@ -1,0 +1,10 @@
+package com.algaworks.algafood.api.model.response;
+
+import lombok.Data;
+
+@Data
+public class ProfileResponse {
+
+    private Long id;
+    private String name;
+}
