+++ src/main/java/com/algaworks/algafood/api/transformer/CityTransformer.java
@@ -1,7 +1,7 @@
 package com.algaworks.algafood.api.transformer;
 
-import com.algaworks.algafood.api.model.CityRequest;
-import com.algaworks.algafood.api.model.CityResponse;
+import com.algaworks.algafood.api.model.request.CityRequest;
+import com.algaworks.algafood.api.model.response.CityResponse;
 import com.algaworks.algafood.domain.model.City;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
