+++ src/main/java/com/algaworks/algafood/api/transformer/PaymentMethodTransformer.java
@@ -1,7 +1,7 @@
 package com.algaworks.algafood.api.transformer;
 
-import com.algaworks.algafood.api.model.PaymentMethodRequest;
-import com.algaworks.algafood.api.model.PaymentMethodResponse;
+import com.algaworks.algafood.api.model.request.PaymentMethodRequest;
+import com.algaworks.algafood.api.model.response.PaymentMethodResponse;
 import com.algaworks.algafood.domain.model.PaymentMethod;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
