+++ src/main/java/com/algaworks/algafood/domain/repository/ProfileRepository.java
@@ -1,0 +1,7 @@
+package com.algaworks.algafood.domain.repository;
+
+import com.algaworks.algafood.domain.model.Profile;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProfileRepository extends JpaRepository<Profile, Long> {
+}
