+++ src/main/java/com/algaworks/algafood/domain/service/RestaurantService.java
@@ -1,7 +1,7 @@
 package com.algaworks.algafood.domain.service;
 
-import com.algaworks.algafood.api.model.RestaurantRequest;
-import com.algaworks.algafood.api.model.RestaurantResponse;
+import com.algaworks.algafood.api.model.request.RestaurantRequest;
+import com.algaworks.algafood.api.model.response.RestaurantResponse;
 import com.algaworks.algafood.api.transformer.RestaurantTransformer;
 import com.algaworks.algafood.domain.exception.*;
 import com.algaworks.algafood.domain.model.Category;
